payment_forms/requirements.txt
django==4.1.1
django-money==3.0
djangorestframework==3.12.4
stripe==4.1.0
python-dotenv==0.21.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
tzdata==2022.2
django-cors-headers==3.13.0